Add base documentation functionality
This adds the following command:

  tachyons src/tachyons-type-scale.css --generate-docs \
           --package=./package.json

This command reads in source CSS and a package to generate
documentation based on templates/readme.md. The templating uses lodash,
and the CSS stats are added to the readme as well.

I've also added a rudimentary test to ensure the docs generation is
working as expected (which is also why I added the tachyons-type-scale
dev dependency).

TODO

* Add immutable-css linting
* Add variables documentation

***

Add basic test and templating functionality

Add package to grab variables

Add io redirection example to docs example